// src/generated/model/klageDtoStatus.ts
/**
 * Generated by orval v6.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type KlageDtoStatus = (typeof KlageDtoStatus)[keyof typeof KlageDtoStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const KlageDtoStatus = {
    SENDT: "SENDT",
    MOTTATT: "MOTTATT",
    UNDER_BEHANDLING: "UNDER_BEHANDLING",
    FERDIG_BEHANDLET: "FERDIG_BEHANDLET",
    HOS_STATSFORVALTER: "HOS_STATSFORVALTER",
} as const;